+++ src/F1_plotting.py
@@ -2,24 +2,17 @@ import logging
 import os
 import textwrap
 
-import matplotlib.pyplot as plt
 import pandas as pd
 import plotly.graph_objs as go
 import plotly.express as px
 
-import networkx as nx
-import oemof.network.graph as graph
+import graphviz
+import oemof
 
 from src.constants import (
-    PATHS_TO_PLOTS,
-    PLOTS_NX,
     PROJECT_DATA,
     LABEL,
     OUTPUT_FOLDER,
-    PATH_OUTPUT_FOLDER,
-    PLOTS_DEMANDS,
-    PLOTS_RESOURCES,
-    LOGFILE,
 )
 
 from src.constants_json_strings import (
@@ -54,9 +47,6 @@ Module F1 describes all the functions that create plots.
 - creating pie chart for cost data
 - creating network graph for the model brackets only working on Ubuntu
 """
-
-
-logging.getLogger("matplotlib.font_manager").disabled = True
 
 
 def convert_plot_data_to_dataframe(plot_data_dict, data_type):
@@ -169,94 +159,247 @@ def parse_simulation_log(path_log_file=None, log_type="ERROR"):
     return logs_dict
 
 
-def draw_graph(
-    dict_values,
-    energysystem,
-    user_input,
-    edge_labels=True,
-    node_color="#eeac7e",
-    edge_color="#eeac7e",
-    show_plot=True,
-    save_plot=True,
-    node_size=5500,
-    with_labels=True,
-    arrows=True,
-    layout="dot",
-):
-    """
+def fixed_width_text(text, char_num=10):
+    """Add linebreaks every char_num characters in a given text.
 
     Parameters
     ----------
-    dict_values: dict
-        Dict with all simulation parameters
-    energysystem:
-    user_input: dict
-    edge_labels: bool
-        Default: True
-    node_color: str
-        Default: "#eeac7e"
-    edge_color: str
-        Default: "#eeac7e"
-    show_plot: bool
-        Default: True
-    save_plot: bool
-        Default: True
-    node_size: int
-        Default: 5500
-    with_labels: bool
-        Default: True
-    arrows: bool
-        Default: True
-    layout: str
-        Default: "dot"
-
+    text: obj:'str'
+        text to apply the linebreaks
+    char_num: obj:'int'
+        max number of characters in a line before a line break
+        Default: 10
     Returns
     -------
+    obj:'str'
+        the text with line breaks after every char_num characters
 
     """
+    # total number of characters in the text
+    text_length = len(text)
+    # integer number of lines of `char_num` character
+    n_lines = int(text_length / char_num)
+    # number of character in the last line
+    last_line_length = text_length % char_num
 
-    # grph = graph.create_nx_graph(
-    #    energysystem, filename=user_input[PATH_OUTPUT_FOLDER] + "network_graph.xml"
-    # )
+    # split the text in lines of `char_num` character
+    split_text = []
+    for i in range(n_lines):
+        split_text.append(text[(i * char_num) : ((i + 1) * char_num)])
 
-    grph = graph.create_nx_graph(energysystem)
+    # I if the last line is not empty
+    if n_lines > 0:
+        if last_line_length > 0:
+            split_text.append(text[((i + 1) * char_num) :])
+        answer = "\n".join(split_text)
+    else:
+        answer = text
+    return answer
 
-    if type(node_color) is dict:
-        node_color = [node_color.get(g, "#AFAFAF") for g in grph.nodes()]
 
-    # set drawing options
-    options = {
-        "prog": "dot",
-        "with_labels": with_labels,
-        "node_color": node_color,
-        "edge_color": edge_color,
-        "node_size": node_size,
-        "arrows": arrows,
-        "font_size": 12,
-        "font_color": "w",
-    }
+class ESGraphRenderer:
+    def __init__(
+        self,
+        energy_system=None,
+        filepath="network",
+        img_format=None,
+        legend=True,
+        txt_width=10,
+        txt_fontsize=10,
+        **kwargs
+    ):
+        """Draw the energy system with Graphviz.
 
-    # draw graph
-    fig, ax = plt.subplots(figsize=(20, 10))
-    pos = nx.drawing.nx_agraph.graphviz_layout(grph, prog=layout)
+        Parameters
+        ----------
+        energy_system: `oemof.solph.network.EnergySystem`
+            The oemof energy stystem
 
-    nx.draw(grph, pos=pos, ax=ax, **options)
+        filepath: str
+            path, where the rendered result shall be saved, if an extension is provided, the format
+            will be automatically adapted except if the `img_format` argument is provided
+            Default: "network"
 
-    # add edge labels for all edges
-    if edge_labels is True and plt:
-        labels = nx.get_edge_attributes(grph, "weight")
-        nx.draw_networkx_edge_labels(grph, pos=pos, edge_labels=labels)
+        img_format: str
+            extension of the available image formats of graphviz (e.g "png", "svg", "pdf", ... )
+            Default: "pdf"
 
-    if show_plot is True:
-        fig.show()
+        legend: bool
+            specify, whether a legend will be added to the graph or not
+            Default: False
 
-    if save_plot is True:
-        path = os.path.join(user_input[PATH_OUTPUT_FOLDER], "network_graph.png")
-        dict_values[PATHS_TO_PLOTS][PLOTS_NX] += [str(path)]
+        txt_width: int
+            max number of characters in a line before a line break
+            Default: 10
 
-        fig.savefig(
-            path, bbox_inches="tight",
+         txt_fontsize: int
+            fontsize of the image's text (components labels)
+            Default: 10
+
+        Returns
+        -------
+        None: render the generated dot graph in the filepath
+        
+        Notes
+        --------
+        When new oemof-solph asset types are added to the available types in the MVS, this function needs to be updated, so that it can render the asset in the graph.
+        """
+        file_name, file_ext = os.path.splitext(filepath)
+
+        if img_format is None:
+            if file_ext != "":
+                img_format = file_ext.replace(".", "")
+            else:
+                img_format = "pdf"
+
+        self.dot = graphviz.Digraph(filename=file_name, format=img_format, **kwargs)
+        self.txt_width = txt_width
+        self.txt_fontsize = str(txt_fontsize)
+        self.busses = []
+
+        if legend is True:
+            with self.dot.subgraph(name="cluster_1") as c:
+                # color of the legend box
+                c.attr(color="black")
+                # title of the legend box
+                c.attr(label="Legends")
+                self.add_bus(subgraph=c)
+                self.add_sink(subgraph=c)
+                self.add_source(subgraph=c)
+                self.add_transformer(subgraph=c)
+                self.add_storage(subgraph=c)
+
+        # draw a node for each of the network's component. The shape depends on the component's type
+        for nd in energy_system.nodes:
+            if isinstance(nd, oemof.solph.network.Bus):
+                self.add_bus(nd.label)
+                # keep the bus reference for drawing edges later
+                self.busses.append(nd)
+            elif isinstance(nd, oemof.solph.network.Sink):
+                self.add_sink(nd.label)
+            elif isinstance(nd, oemof.solph.network.Source):
+                self.add_source(nd.label)
+            elif isinstance(nd, oemof.solph.network.Transformer):
+                self.add_transformer(nd.label)
+            elif isinstance(nd, oemof.solph.components.GenericStorage):
+                self.add_storage(nd.label)
+            else:
+                logging.warning(
+                    "The component {} of type {} is not implemented in the rendering "
+                    "function of the energy model network graph drawer. It will be "
+                    "rendered as an ellipse".format(nd.label, type(nd))
+                )
+                self.add_component(nd.label)
+
+        # draw the edges between the nodes based on each bus inputs/outputs
+        for bus in self.busses:
+            for component in bus.inputs:
+                # draw an arrow from the component to the bus
+                self.connect(component, bus)
+            for component in bus.outputs:
+                # draw an arrow from the bus to the component
+                self.connect(bus, component)
+
+    def add_bus(self, label="Bus", subgraph=None):
+        if subgraph is None:
+            dot = self.dot
+        else:
+            dot = subgraph
+        dot.node(
+            label,
+            shape="rectangle",
+            fontsize="10",
+            fixedsize="shape",
+            width="4.1",
+            height="0.3",
+            style="filled",
+            color="lightgrey",
         )
+
+    def add_sink(self, label="Sink", subgraph=None):
+        if subgraph is None:
+            dot = self.dot
+        else:
+            dot = subgraph
+        dot.node(
+            fixed_width_text(label, char_num=self.txt_width),
+            shape="trapezium",
+            fontsize=self.txt_fontsize,
+        )
+
+    def add_source(self, label="Source", subgraph=None):
+        if subgraph is None:
+            dot = self.dot
+        else:
+            dot = subgraph
+        dot.node(
+            fixed_width_text(label, char_num=self.txt_width),
+            shape="invtrapezium",
+            fontsize=self.txt_fontsize,
+        )
+
+    def add_transformer(self, label="Transformer", subgraph=None):
+        if subgraph is None:
+            dot = self.dot
+        else:
+            dot = subgraph
+        dot.node(
+            fixed_width_text(label, char_num=self.txt_width),
+            shape="rectangle",
+            fontsize=self.txt_fontsize,
+        )
+
+    def add_storage(self, label="Storage", subgraph=None):
+        if subgraph is None:
+            dot = self.dot
+        else:
+            dot = subgraph
+        dot.node(
+            fixed_width_text(label, char_num=self.txt_width),
+            shape="rectangle",
+            style="rounded",
+            fontsize=self.txt_fontsize,
+        )
+
+    def add_component(self, label="component", subgraph=None):
+        if subgraph is None:
+            dot = self.dot
+        else:
+            dot = subgraph
+        dot.node(
+            fixed_width_text(label, char_num=self.txt_width),
+            fontsize=self.txt_fontsize,
+        )
+
+    def connect(self, a, b):
+        """Draw an arrow from node a to node b
+
+        Parameters
+        ----------
+        a: `oemof.solph.network.Node`
+            An oemof node (usually a Bus or a Component)
+
+        b: `oemof.solph.network.Node`
+            An oemof node (usually a Bus or a Component)
+        """
+        if not isinstance(a, oemof.solph.network.Bus):
+            a = fixed_width_text(a.label, char_num=self.txt_width)
+        else:
+            a = a.label
+        if not isinstance(b, oemof.solph.network.Bus):
+            b = fixed_width_text(b.label, char_num=self.txt_width)
+        else:
+            b = b.label
+
+        self.dot.edge(a, b)
+
+    def view(self, **kwargs):
+        """Call the view method of the DiGraph instance"""
+        self.dot.view(**kwargs)
+
+    def render(self, **kwargs):
+        """Call the render method of the DiGraph instance"""
+        self.dot.render(**kwargs)
 
 
 def get_color(idx_line, color_list=None):